requirements.txt
fastapi
uvicorn==0.38.0
httpx==0.28.1
pydantic==2.12.4
python-dotenv==1.2.1
starlette==0.49.3
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
h11==0.16.0
httpcore==1.0.9
idna==3.11
annotated-doc==0.0.4
annotated-types==0.7.0
pydantic_core==2.41.5
setuptools==66.1.1
sniffio==1.3.1
typing_extensions==4.15.0
typing-inspection==0.4.2
mcp==1.21.2
fastmcp
fqdn==1.5.0